==> evaluation/3_evaluate_generated.ipynb <==
from uuid import uuid4

import pandas as pd
from langchain.chat_models import init_chat_model

from plum_chatbot.configs.settings import Settings
from plum_chatbot.datasources.qdrant_datasource import QdrantDatasource
from plum_chatbot.di_containers.datasources_containers import Container, container
from plum_chatbot.schemas.schema import UserInput

settings = Settings()

judge_llm = init_chat_model(
    "llama3.2",  # qwen3:8b
    model_provider="ollama",
    base_url="host.docker.internal",
    temperature=0.5,
)

golden_set = pd.read_csv("data/generated_golden_set_with_answers.csv")

golden_set

container.init_resources()
for k, v in Container.providers.items():
    provider = v()
    if isinstance(provider, QdrantDatasource):
        provider.setup()

from langsmith import Client

client = Client(api_key=settings.LANGSMITH_API_KEY)

# Programmatically create a dataset in LangSmith
# For other dataset creation methods, see:
# https://docs.smith.langchain.com/evaluation/how_to_guides/manage_datasets_programmatically
# https://docs.smith.langchain.com/evaluation/how_to_guides/manage_datasets_in_application
dataset = client.create_dataset(
    dataset_name="Generated FAQ dataset",
    description="A sample dataset generated from LLM (llama 3.2).",
)

# Create examples
examples = [
    {
        "inputs": {"question": row["question"]},
        "outputs": {"answer": row["real_answer"]},
        "metadata": {"id": row["id"], "content_used": row["content_used"]},
    }
    for _, row in golden_set.iterrows()
]

# Add examples to the dataset
client.create_examples(dataset_id=dataset.id, examples=examples);

async def correct(outputs: dict, reference_outputs: dict) -> bool:
    instructions = (
        "Given an actual answer and an expected answer, determine whether"
        " the actual answer contains all of the information in the"
        " expected answer. Respond with 'CORRECT' if the actual answer"
        " does contain all of the expected information and 'INCORRECT'"
        " otherwise. Do not include anything else in your response."
    )
    # Our graph outputs a State dictionary, which in this case means
    # we'll have a 'messages' key and the final message should
    # be our actual answer.
    actual_answer = outputs["messages"][-1].content
    expected_answer = reference_outputs["answer"]
    user_msg = f"ACTUAL ANSWER: {actual_answer}\n\nEXPECTED ANSWER: {expected_answer}"
    response = await judge_llm.ainvoke(
        [
            {"role": "system", "content": instructions},
            {"role": "user", "content": user_msg},
        ]
    )
    return response.content.upper() == "CORRECT"

from plum_chatbot.agents.agents import _handle_input
from plum_chatbot.agents.my_agent import RagAgent


def example_to_state(inputs: dict) -> dict:
    return {"messages": [{"role": "user", "content": inputs["question"]}]}


async def to_agent_input(inputs: dict) -> dict:
    kwargs, run_id, thread_id = await _handle_input(
        UserInput(message=inputs["question"], thread_id=str(uuid4())), agent.graph
    )
    return kwargs


#     input = {"messages": [HumanMessage(content=inputs["question"])]}
#     run_id = uuid4()
#     thread_id = str(uuid4())
#     user_id = str(uuid4())
#     configurable = {
#         "thread_id": thread_id,
#         "model": "rag",
#         "user_id": user_id,
#     }
#     config = RunnableConfig(
#         configurable=configurable,
#         run_id=run_id,
#     )
#     kwargs = {
#         "input": input,
#         "config": config,
#     }
#     return kwargs


agent = RagAgent()


# We use LCEL declarative syntax here.
# Remember that langgraph graphs are also langchain runnables.
# Compose the target as a function that unpacks the dict returned by to_agent_input
async def target(inputs):
    kwargs = await to_agent_input(inputs)
    return await agent.graph.ainvoke(**kwargs)

from langsmith import aevaluate

experiment_results = await aevaluate(
    target,
    data="Generated FAQ dataset",
    evaluators=[correct],
    max_concurrency=4,  # optional
    experiment_prefix="llama3.2",  # optional
)
